Create twosum leet code
right now, my approach to the problem is to brute force the solution. What I planned is to find every combination that matches the target. Using for loops. But my problem here is my syntax since i came from c and c++. I'm still used to a lower-level type of coding so i did i = 0. and not very familiar with the new syntax. But it seems my logic should be good.