Final verison twosum leet code
This should be the most elegant and efficient version of the code. It's called HashMap and it takes advantages of storing values and its keys. Saying has i seen this number before, then it pairs it well. A more memory and resource saver. Learned enumerate a way to replace the len().