Create palindrome test
this is my first attempt at this. My logic here is to check the first array and last array and to match it. I will also check if there is a middle, if there is no middle it is false. Right now, I got it but I haven't accounted for the numbers between the middle and beginning- end.