Update roman numerals to integer
this is able to shorten the index and doesn't need to rely on putting the special characters in it. Instead, this one recognizes the if the character is less than the other one making it a subtractive case and adjust and skips it after adding the value. This the most efficient and optimize solution I can think of currently.